Moved logger function to helpers.py
