+++ move_researchdata/wrapper.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+
+import click
+
+def look_for_runs(root_path):
+    found_paths = glob.glob(os.path.join(root_path, '*'))
+    regex = '^[0-9]{6}_A[0-9]{5}_[0-9]{4}_.{10}$'
+    return [path for path in found_paths if re.search(regex, os.path.basename(path))]
